% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{accept_reject}
\alias{accept_reject}
\title{Accept or Reject Samples}
\usage{
accept_reject(sample_size, linear_part, offset)
}
\arguments{
\item{sample_size}{Number of samples generated}

\item{linear_part}{Linear part of the constraint}

\item{offset}{Offset part of the constraint}
}
\value{
Random samples that satisfy the box constraint.
}
\description{
Random samples from normal distribution that satisfy the box constraint.
}
\examples{
# Fit the group lasso optimization model
Z = matrix(rnorm(10*3), nrow = 10, ncol = 3) +
    matrix(replicate(3, matrix(0, nrow = 10, ncol = 1)),
    nrow = 10)
errorTerm = MASS::mvrnorm(n=10, mu=rep(0,2),
            Sigma=rbind(c(1, 0.8),c(0.8, 1)))
D = Z \%*\% rep(1, 3) + errorTerm[1:10, 2]
Y = D * 1 + errorTerm[1:10, 1]
gl <- group_lasso_iv(Y, D, Z)
model <- fit_tsls(gl)

# Testing samples
white_con = whiten(model$sampler$affine_con)
accept_reject(100,
              white_con$linear_part,
              white_con$offset)
}
\references{
Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2019). Inference After Selecting
Plausibly Valid Instruments with Application to Mendelian Randomization.

Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2020). Inferring Treatment
Effects After Testing Instrument Strength in Linear Models.
}
