% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_lasso_iv.R
\name{bigaussian_instance}
\alias{bigaussian_instance}
\title{Bigaussian Instance}
\usage{
bigaussian_instance(
  n = 1000,
  p = 10,
  s = 3,
  snr = 7,
  gsnr = 1,
  beta = 1,
  Sigma = rbind(c(1, 0.8), c(0.8, 1)),
  rho = 0,
  scale = FALSE,
  center = TRUE
)
}
\arguments{
\item{n}{Number of samples.}

\item{p}{Number of instruments.}

\item{s}{Number of invalid instruments.}

\item{snr}{Coefficient of invalid instruments.}

\item{gsnr}{True gamma parameter.}

\item{beta}{True beta parameter.}

\item{Sigma}{Noise variance-covariance matrix.}

\item{rho}{Z matrix structure.}

\item{scale}{Binary indicator of whether it is a weak IV case.}

\item{center}{Binary indicator of whether re-center the distribution to 0.}
}
\value{
A bigaussian instance, which is a list containing
\item{Y}{The outcome.}
\item{D}{The treatment.}
\item{Z}{The p instruments.}
\item{beta}{True beta parameter.}
\item{gamma}{A vector of of length p of true gamma parameter.}
}
\description{
Generate a dataset, including the outcome, treatment, and instrument matrices,
for simulation. The dataset can be used to test the instrument strength and
the treatment effect.
}
\examples{
# True beta value is 1
bigaussian_instance(n=10, p=5, s=2, snr=3, gsnr=1, beta=1,
Sigma = rbind(c(1, 0.8), c(0.8, 1)), rho=0)

# True beta value is 0
bigaussian_instance(n=100, p=10, s=7, snr=3, gsnr=1, beta=0,
Sigma = rbind(c(1, 0.8), c(0.8, 1)), rho=0)
}
\references{
Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2019). Inference After Selecting
Plausibly Valid Instruments with Application to Mendelian Randomization.

Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2020). Inferring Treatment
Effects After Testing Instrument Strength in Linear Models.
}
\seealso{
\link[MASS]{mvrnorm} for sampling from multivariate normal distribution.
}
