% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{coefficient_pvalues_iv}
\alias{coefficient_pvalues_iv}
\title{Coefficient pvalues with IV for TSLS Test Statistic}
\usage{
coefficient_pvalues_iv(
  observed_target,
  target_cov,
  score_cov,
  opt_sampler,
  parameter = NULL,
  sample_args = NULL,
  sample,
  alternatives = NULL
)
}
\arguments{
\item{observed_target}{A vector of parameters, representing coordinates of the target}

\item{target_cov}{Covariance of the target (e.g. TSLS)}

\item{score_cov}{Covariance of the target score (e.g. TSLS score)}

\item{opt_sampler}{A sampler generated from the optimization problem}

\item{parameter}{A vector of parameters at which to evaluate p-values. Defaults
to the zeros matrix}

\item{sample_args}{Default to NULL}

\item{sample}{Represents a sample of the target. Allows reuse of the same sample
for construction of confidence intervals, hypothesis tests, etc.}

\item{alternatives}{A list of ['greater', 'less', 'twosided'],
which indicates what alternative to use}
}
\value{
Sampled p-values.
}
\description{
Construct selective p-values for each parameter of the target. The p-values
are computed for the optimization problem with instrumental variables.
We compute the p-values of the Two-Stage Least-Squares (TSLS) test statistic.
}
\examples{
# Fit the group lasso optimization model
Z = matrix(rnorm(10*3), nrow = 10, ncol = 3) +
    matrix(replicate(3, matrix(0, nrow = 10, ncol = 1)),
    nrow = 10)
errorTerm = MASS::mvrnorm(n=10, mu=rep(0,2),
            Sigma=rbind(c(1, 0.8),c(0.8, 1)))
D = Z \%*\% rep(1, 3) + errorTerm[1:10, 2]
Y = D * 1 + errorTerm[1:10, 1]
gl <- group_lasso_iv(Y, D, Z)
model <- fit_tsls(gl)

# Estimate covariance and obtain summary statistics
cov = estimate_covariance(Y, D, Z)
s <- summary_tsls(gl, model$sampler,
                  Sigma_11 = cov[1,1], Sigma_12 = cov[1,2])

# Compute p-values for TSLS test statistic
pivots = coefficient_pvalues_iv(observed_target=s$observed_target,
                                target_cov=s$cov_target,
                                score_cov=s$cov_target_score,
                                opt_sampler=model$sampler,
                                parameter=0,
                                sample=s$opt_sample,
                                alternatives=rep('twosided',1))
pivots
}
\references{
Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2019). Inference After Selecting
Plausibly Valid Instruments with Application to Mendelian Randomization.

Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2020). Inferring Treatment
Effects After Testing Instrument Strength in Linear Models.
}
\seealso{
\link{coefficient_pvalues_iv_ar} that computes selective p-values for AR
statistic. The method to compute p-values for TSLS and AR statistics are similar.

\link{optimization_intervals} to initialize an interval for the optimization
problem, which specify the observed target and its covariance as well as
optimization samples and sampler. We need to initialize an optimization_intervals
before computing p-values.

\link{pivot_iv} for calculating the pivot quantities of the optimization
intervals before computing p-values.
}
