% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals_ar.R
\name{confidence_interval_coverage}
\alias{confidence_interval_coverage}
\title{Test Coverage of Confidence Interval for AR Test Statistic}
\usage{
confidence_interval_coverage(
  observed_target,
  target_cov,
  score_cov,
  linear_func,
  gl_ar,
  opt_sampler,
  beta_reference,
  beta_candidate,
  sample_args = NULL,
  sample,
  level = 0.95
)
}
\arguments{
\item{observed_target}{A vector of parameters, representing coordinates of the target}

\item{target_cov}{Covariance of the target (e.g. AR)}

\item{score_cov}{Covariance of the target score (e.g. AR score)}

\item{linear_func}{Linear function}

\item{gl_ar}{The group lasso object for AR test statistic representing the
penalized convex optimization equation (generated by
group_lasso_iv function)}

\item{opt_sampler}{Optimization sampler}

\item{beta_reference}{Reference value of beta parameter}

\item{beta_candidate}{Candidate value of beta parameter}

\item{sample_args}{Default to NULL (optional)}

\item{sample}{Represents a sample of the target. Allows reuse of the same sample
for construction of confidence intervals, hypothesis tests, etc.}

\item{level}{Specify the confidence level (optional)}
}
\value{
TRUE if p-value is higher than the significance level (i.e. not
        significant) and FALSE otherwise
}
\description{
Test selective confidence interval for AR test statistic with
instrumental variables.
}
\examples{
# Fit the group lasso optimization model
Z = matrix(rnorm(10*3), nrow = 10, ncol = 3) +
    matrix(replicate(3, matrix(0, nrow = 10, ncol = 1)),
    nrow = 10)
errorTerm = MASS::mvrnorm(n=10, mu=rep(0,2),
            Sigma=rbind(c(1, 0.8),c(0.8, 1)))
D = Z \%*\% rep(1, 3) + errorTerm[1:10, 2]
Y = D * 1 + errorTerm[1:10, 1]
gl_ar <- group_lasso_iv_ar(Y, D, Z)
model_ar <- fit_ar(gl_ar)

# Estimate covariance and obtain summary statistics
cov = estimate_covariance(Y, D, Z)
s <- summary_ar(gl_ar, model_ar$sampler,
                Sigma_11=cov[1,1], Sigma_12=cov[1,2])

# Test confidence interval coverage for AR test statistic
confidence_interval_coverage(observed_target=s$observed_target_tsls,
                             target_cov=s$cov_target,
                             score_cov=s$cov_target_score,
                             linear_func=s$K2,
                             gl_ar=gl_ar,
                             opt_sampler=model_ar$sampler,
                             beta_reference=s$two_stage_ls[1,1],
                             beta_candidate=0,
                             sample=s$opt_sample,
                             level=0.95)
}
\references{
Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2019). Inference After Selecting
Plausibly Valid Instruments with Application to Mendelian Randomization.

Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2020). Inferring Treatment
Effects After Testing Instrument Strength in Linear Models.
}
\seealso{
\link{optimization_intervals} to initialize an interval for the optimization
problem, which specify the observed target and its covariance as well as
optimization samples and sampler. We need to initialize an optimization_intervals
before computing confidence intervals.

\link{pivot_iv_ar} for calculating the pivot quantities of the optimization
intervals before computing p-values.

\link{confidence_intervals_iv_ar} for constructing selective confidence
intervals for AR test statistic.

\link{confidence_interval_iv_ar} for constructing a selective
confidence interval for a specific parameter of the target for AR
test statistic.
}
