% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{confidence_interval_iv}
\alias{confidence_interval_iv}
\title{Confidence Interval}
\usage{
confidence_interval_iv(
  intervals,
  linear_func,
  level = 0.9,
  how_many_sd = 20,
  guess = NULL
)
}
\arguments{
\item{intervals}{Optimization intervals}

\item{linear_func}{Linear function}

\item{level}{Specify the confidence level (optional)}

\item{how_many_sd}{Specify the number of standard deviations (optional)}

\item{guess}{Initial guess at upper and lower limit (optional)}
}
\value{
A confidence interval object, which is a list of:
\item{upper}{Upper limit of the confidence interval}
\item{lower}{Lower limit of the confidence interval}
}
\description{
Construct a selective confidence interval for a specific observed target
statistic.
}
\examples{
# Fit the group lasso optimization model
Z = matrix(rnorm(10*3), nrow = 10, ncol = 3) +
    matrix(replicate(3, matrix(0, nrow = 10, ncol = 1)),
    nrow = 10)
errorTerm = MASS::mvrnorm(n=10, mu=rep(0,2),
            Sigma=rbind(c(1, 0.8),c(0.8, 1)))
D = Z \%*\% rep(1, 3) + errorTerm[1:10, 2]
Y = D * 1 + errorTerm[1:10, 1]
gl <- group_lasso_iv(Y, D, Z)
model <- fit_tsls(gl)

# Estimate covariance and obtain summary statistics
cov = estimate_covariance(Y, D, Z)
s <- summary_tsls(gl, model$sampler,
                  Sigma_11 = cov[1,1], Sigma_12 = cov[1,2])

# Construct the optimization intervals before computing confidence intervals
# for TSLS test statistic
ndraw = nrow(s$opt_sample)
opt_sampling_info = list(sets::tuple(model$sampler,
                                     s$opt_sample,
                                     s$cov_target,
                                     s$cov_target_score))
intervals <- optimization_intervals(opt_sampling_info,
                                    s$observed_target,
                                    ndraw)

# Compute the confidence interval for a specific observed target
keep = matrix(0, nrow(s$observed_target), ncol(s$observed_target))
keep[1] = 1
confidence_interval_iv(intervals, keep, level=0.95)
confidence_interval_iv(intervals, keep, level=0.95,
                       guess=c(0.5, 1.5))
}
\references{
Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2019). Inference After Selecting
Plausibly Valid Instruments with Application to Mendelian Randomization.

Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2020). Inferring Treatment
Effects After Testing Instrument Strength in Linear Models.
}
\seealso{
\link{optimization_intervals} to initialize an interval for the optimization
problem, which specify the observed target and its covariance as well as
optimization samples and sampler. We need to initialize an optimization_intervals
before computing confidence intervals.

\link{confidence_intervals_iv} for constructing selective confidence
intervals for TSLS test statistic.

\link{confidence_intervals_iv_ar} for constructing selective confidence
intervals for AR test statistic.

\link{confidence_interval_iv_ar} for constructing a selective
confidence interval for a specific parameter of the target for AR
test statistic.

\link{pivot_iv} for calculating the pivot quantities/p-values of the
optimization intervals for TSLS test statistic.
}
