% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{confidence_intervals_iv}
\alias{confidence_intervals_iv}
\title{Confidence Intervals}
\usage{
confidence_intervals_iv(
  observed_target,
  target_cov,
  score_cov,
  sample_args = NULL,
  opt_sampler,
  sample,
  level = 0.9,
  initial_guess = NULL
)
}
\arguments{
\item{observed_target}{A vector of parameters, representing coordinates of the target}

\item{target_cov}{Covariance of the target (e.g. TSLS)}

\item{score_cov}{Covariance of the target score (e.g. TSLS score)}

\item{sample_args}{Default to NULL (optional)}

\item{opt_sampler}{Optimization sampler}

\item{sample}{Represents a sample of the target. Allows reuse of the same sample
for construction of confidence intervals, hypothesis tests, etc.}

\item{level}{Specify the confidence level (optional)}

\item{initial_guess}{Initial guesses at upper and lower limits (optional)}
}
\value{
List of confidence intervals.
}
\description{
Construct selective confidence intervals for each observed target.
The confidence intervals are computed for the optimization problem with
instrumental variables. We compute the confidence intervals of the
Two-Stage Least-Squares (TSLS) test statistic.
}
\examples{
# Fit the group lasso optimization model
Z = matrix(rnorm(10*3), nrow = 10, ncol = 3) +
    matrix(replicate(3, matrix(0, nrow = 10, ncol = 1)),
    nrow = 10)
errorTerm = MASS::mvrnorm(n=10, mu=rep(0,2),
            Sigma=rbind(c(1, 0.8),c(0.8, 1)))
D = Z \%*\% rep(1, 3) + errorTerm[1:10, 2]
Y = D * 1 + errorTerm[1:10, 1]
gl <- group_lasso_iv(Y, D, Z)
model <- fit_tsls(gl)

# Estimate covariance and obtain summary statistics
cov = estimate_covariance(Y, D, Z)
s <- summary_tsls(gl, model$sampler,
                  Sigma_11 = cov[1,1], Sigma_12 = cov[1,2])

# Compute confidence intervals for TSLS test statistic
confidence_intervals_iv(observed_target=s$observed_target,
                        target_cov=s$cov_target,
                        score_cov=s$cov_target_score,
                        opt_sampler=model$sampler,
                        sample=s$opt_sample,
                        level=0.95)
}
\references{
Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2019). Inference After Selecting
Plausibly Valid Instruments with Application to Mendelian Randomization.

Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2020). Inferring Treatment
Effects After Testing Instrument Strength in Linear Models.
}
\seealso{
\link{optimization_intervals} to initialize an interval for the optimization
problem, which specify the observed target and its covariance as well as
optimization samples and sampler. We need to initialize an optimization_intervals
before computing confidence intervals.

\link{confidence_interval_iv} for constructing a selective confidence
intervals for a specific parameter of the target.

\link{confidence_intervals_iv_ar} for constructing selective confidence
intervals for AR test statistic.

\link{confidence_interval_iv_ar} for constructing a selective
confidence interval for a specific parameter of the target for AR
test statistic.
}
