% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{covariance_factors}
\alias{covariance_factors}
\title{Covariance Factors}
\usage{
covariance_factors(con, force = TRUE)
}
\arguments{
\item{con}{Constraints}

\item{force}{If True, force a recomputation of the covariance.
If not, assumes that covariance has not changed.}
}
\value{
A covariance_factors object, which is a list of
\item{sqrt_cov}{Squared root of the covariance matrix}
\item{sqrt_inv}{Squared root of the inverse covariance matrix}
\item{rowspace}{Row space matrix}
}
\description{
Finding a possibly non-square square-root.
}
\examples{
# Fit the group lasso optimization model
Z = matrix(rnorm(10*3), nrow = 10, ncol = 3) +
    matrix(replicate(3, matrix(0, nrow = 10, ncol = 1)),
    nrow = 10)
errorTerm = MASS::mvrnorm(n=10, mu=rep(0,2),
            Sigma=rbind(c(1, 0.8),c(0.8, 1)))
D = Z \%*\% rep(1, 3) + errorTerm[1:10, 2]
Y = D * 1 + errorTerm[1:10, 1]
gl <- group_lasso_iv(Y, D, Z)
model <- fit_tsls(gl)

# Covariance factors for the original box constraint
covariance_factors(model$sampler$affine_con)
# still compute sqrt_cov, sqrt_inv, rowspace as the constraint
# doesn't contain such attributes
covariance_factors(model$sampler$affine_con, force=FALSE)

# Whiten the box constraint
white_con = whiten(model$sampler$affine_con)

# Covariance factors for the white box constraint
covariance_factors(white_con)
covariance_factors(white_con, force=FALSE)
}
\references{
Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2019). Inference After Selecting
Plausibly Valid Instruments with Application to Mendelian Randomization.

Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2020). Inferring Treatment
Effects After Testing Instrument Strength in Linear Models.
}
