% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_lasso_iv.R
\name{estimate_covariance}
\alias{estimate_covariance}
\title{Estimate Covariance}
\usage{
estimate_covariance(Y, D, Z)
}
\arguments{
\item{Y}{The outcome.}

\item{D}{The treatment.}

\item{Z}{The p instruments.}
}
\value{
The estimated 2x2 covariance matrix.
}
\description{
Compute the covariance matrix of TSLS or AR test statistic.
}
\examples{
# Estimate the covariance
Z = matrix(rnorm(10*3), nrow = 10, ncol = 3) +
    matrix(replicate(3, matrix(0, nrow = 10, ncol = 1)),
    nrow = 10)
errorTerm = MASS::mvrnorm(n=10, mu=rep(0,2),
           Sigma=rbind(c(1, 0.8),c(0.8, 1)))
D = Z \%*\% rep(1, 3) + errorTerm[1:10, 2]
Y = D * 1 + errorTerm[1:10, 1]
estimate_covariance(Y, D, Z)
}
\references{
Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2019). Inference After Selecting
Plausibly Valid Instruments with Application to Mendelian Randomization.

Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2020). Inferring Treatment
Effects After Testing Instrument Strength in Linear Models.
}
