% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{forward_map}
\alias{forward_map}
\title{Forward Map}
\usage{
forward_map(W, white_con)
}
\arguments{
\item{W}{The input matrix}

\item{white_con}{Whiten constraints}
}
\value{
The forward map of white_con.
}
\description{
A callable forward_map.
}
\examples{
# Fit the group lasso optimization model
Z = matrix(rnorm(10*3), nrow = 10, ncol = 3) +
    matrix(replicate(3, matrix(0, nrow = 10, ncol = 1)),
    nrow = 10)
errorTerm = MASS::mvrnorm(n=10, mu=rep(0,2),
            Sigma=rbind(c(1, 0.8),c(0.8, 1)))
D = Z \%*\% rep(1, 3) + errorTerm[1:10, 2]
Y = D * 1 + errorTerm[1:10, 1]
gl <- group_lasso_iv(Y, D, Z)
model <- fit_tsls(gl)

# Whiten constraint and create forward_map
white_con = whiten(model$sampler$affine_con)
forward_map(model$sampler$initial_point, white_con)
}
\references{
Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2019). Inference After Selecting
Plausibly Valid Instruments with Application to Mendelian Randomization.

Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2020). Inferring Treatment
Effects After Testing Instrument Strength in Linear Models.
}
