% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_lasso_iv.R
\name{group_lasso_iv}
\alias{group_lasso_iv}
\title{Group Lasso Model with Instrumental Variables}
\usage{
group_lasso_iv(
  Y,
  D,
  Z,
  penalty = NULL,
  ridge_term = 0,
  randomizer_scale = NULL,
  perturb = NULL,
  C0 = 10
)
}
\arguments{
\item{Y}{The outcome.}

\item{D}{The treatment.}

\item{Z}{The p instruments.}

\item{penalty}{Lambda term in the optimization model.}

\item{ridge_term}{L2 regularization term (if applicable). Default to 0.}

\item{randomizer_scale}{Standard deviation of random noise distribution.}

\item{perturb}{Initial value of the randomization term (i.e. omega).}

\item{C0}{Threshold of pre-test on instrument strength. Default to 10.}
}
\value{
A group lasso object, which is a list of:
\item{data_part}{The S matrix in equation.}
\item{loglike}{Goodness of fit of data_part. Regress data_part on a zeros
vector that represent the intercept and an identity matrix that represents
the independent variables.}
\item{penalty}{Lambda term; the weights of penalty. If penalty is not specified,
the method calculates the penalty term based on the equation:
\eqn{\lambda = \sqrt{C_0 \frac{p}{n-p}\left(\sum_{i=1}^{n}(D_i-Z_i^T\hat{\gamma})^2\right)}}}
\item{randomizer_scale}{Standard deviation of random noise.}
\item{Y}{The outcome.}
\item{D}{The treatment.}
\item{Z}{The p instruments.}
}
\description{
Initialize outcome, treatment, and instrument matrices as well as
other terms for penalized convex optimization program.
}
\details{
This function sets up the penalized convex optimization model:

\eqn{\hat{v} = \underset{v}{\operatorname{argmin}} \frac{1}{2}||v-S||_2^2 + \lambda||v||_2 - \omega^Tv}

where S is the data_part, \eqn{\lambda} is the weight of penalty, and \eqn{\omega}
is a random noise from a known Normal distribution.
}
\examples{
# Initialize the group_lasso_iv object with other term set to default
Z = matrix(rnorm(10*3), nrow = 10, ncol = 3) +
    matrix(replicate(3, matrix(0, nrow = 10, ncol = 1)),
    nrow = 10)
errorTerm = MASS::mvrnorm(n=10, mu=rep(0,2),
            Sigma=rbind(c(1, 0.8),c(0.8, 1)))
D = Z \%*\% rep(1, 3) + errorTerm[1:10, 2]
Y = D * 1 + errorTerm[1:10, 1]
group_lasso_iv(Y, D, Z)

# Initalize the group_lasso_iv object specifying other terms
group_lasso_iv(Y, D, Z, C0=10)
group_lasso_iv(Y, D, Z, penalty=5, randomizer_scale=0.8, C0=10)
}
\references{
Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2019). Inference After Selecting
Plausibly Valid Instruments with Application to Mendelian Randomization.

Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2020). Inferring Treatment
Effects After Testing Instrument Strength in Linear Models.
}
\seealso{
\link{group_lasso_iv_ar} that fits a similar group lasso object. It
initializes the group lasso model with instrumental variables for an
AR test statistic.

\link[stats]{glm} for fitting generalized linear models
}
