% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_lasso_iv_ar.R
\name{group_lasso_iv_ar}
\alias{group_lasso_iv_ar}
\title{Group Lasso Model with Instrumental Variables for AR Statistic}
\usage{
group_lasso_iv_ar(
  Y,
  D,
  Z,
  penalty = NULL,
  ridge_term = NULL,
  randomizer_scale = NULL,
  perturb = NULL,
  C0 = NULL
)
}
\arguments{
\item{Y}{The outcome}

\item{D}{The treatment}

\item{Z}{The p instruments}

\item{penalty}{Lambda term in equation (13) and (15)}

\item{ridge_term}{L2 regularization term (if applicable)}

\item{randomizer_scale}{Standard deviation of random noise distribution}

\item{perturb}{Initial value of the randomization term (i.e. omega)}

\item{C0}{Threshold of pre-test on instrument strength}
}
\value{
A group lasso object, which is a list of:
\item{data_part}{The S matrix in equation}
\item{loglike}{Goodness of fit of data_part}
\item{penalty}{Lambda term; the weights of penalty}
\item{randomizer_scale}{Standard deviation of random noise}
\item{Y}{The outcome}
\item{D}{The treatment}
\item{Z}{The p instruments}
}
\description{
Initialize outcome, treatment, and instrument matrices as well as
other terms for penalized convex optimization program.
}
\examples{
# Initialize the group_lasso_iv_ar object with other term set to default
Z = matrix(rnorm(10*3), nrow = 10, ncol = 3) +
    matrix(replicate(3, matrix(0, nrow = 10, ncol = 1)),
    nrow = 10)
errorTerm = MASS::mvrnorm(n=10, mu=rep(0,2),
            Sigma=rbind(c(1, 0.8),c(0.8, 1)))
D = Z \%*\% rep(1, 3) + errorTerm[1:10, 2]
Y = D * 1 + errorTerm[1:10, 1]
group_lasso_iv_ar(Y, D, Z)
}
\references{
Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2019). Inference After Selecting
Plausibly Valid Instruments with Application to Mendelian Randomization.

Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2020). Inferring Treatment
Effects After Testing Instrument Strength in Linear Models.
}
\seealso{
\link{group_lasso_iv} that fits a similar group lasso object. It
initializes the group lasso model with instrumental variables for an
TSLS test statistic.

\link[stats]{glm} for fitting generalized linear models.
}
