% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{inverse_map}
\alias{inverse_map}
\title{Inverse Map}
\usage{
inverse_map(Z, white_con)
}
\arguments{
\item{Z}{The p instruments}

\item{white_con}{Whiten constraints}
}
\value{
The inverse map of white_con.
}
\description{
A callable inverse_map.
}
\examples{
# Fit the group lasso optimization model
Z = matrix(rnorm(10*3), nrow = 10, ncol = 3) +
    matrix(replicate(3, matrix(0, nrow = 10, ncol = 1)),
    nrow = 10)
errorTerm = MASS::mvrnorm(n=10, mu=rep(0,2),
            Sigma=rbind(c(1, 0.8),c(0.8, 1)))
D = Z \%*\% rep(1, 3) + errorTerm[1:10, 2]
Y = D * 1 + errorTerm[1:10, 1]
gl <- group_lasso_iv(Y, D, Z)
model <- fit_tsls(gl)

# Whiten the box constraint
white_con = whiten(model$sampler$affine_con)
white_Y = forward_map(model$sampler$initial_point, white_con)
Y = as.matrix(model$sampler$initial_point)
direction_of_interest = rnorm(length(Y))
white_direction_of_interest =
forward_map(model$sampler$affine_con$covariance \%*\% direction_of_interest,
white_con)

# Generate white samples from truncated normal distribution
white_samples = sample_truncnorm_white(white_con$linear_part,
white_con$offset, white_Y, white_direction_of_interest,
how_often=2000, ndraw=1000, burnin=1000, sigma=1,
use_constraint_directions=TRUE, use_random_directions=TRUE)

# Construct inverse map of white samples
inverse_map(white_samples, white_con)
}
\references{
Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2019). Inference After Selecting
Plausibly Valid Instruments with Application to Mendelian Randomization.

Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2020). Inferring Treatment
Effects After Testing Instrument Strength in Linear Models.
}
