% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_lasso_iv.R
\name{log_density}
\alias{log_density}
\title{Log density}
\usage{
log_density(
  logdens_linear,
  offset,
  cond_prec,
  score,
  opt,
  active_directions,
  lagrange,
  Z
)
}
\arguments{
\item{logdens_linear}{Linear part of log density}

\item{offset}{Offset part of log density}

\item{cond_prec}{Conditional precision}

\item{score}{Score state for opt variables}

\item{opt}{State for opt variables}

\item{active_directions}{Active directions of vector v (same as opt_linear)}

\item{lagrange}{Lambda term; the weights of penalty}

\item{Z}{The p instruments}
}
\value{
The log density of given equation.
}
\description{
Calculate the log density of given equation. It includes the Jacobian term.
}
\examples{
# Fit the group lasso optimization model
Z = matrix(rnorm(10*3), nrow = 10, ncol = 3) +
    matrix(replicate(3, matrix(0, nrow = 10, ncol = 1)),
    nrow = 10)
errorTerm = MASS::mvrnorm(n=10, mu=rep(0,2),
            Sigma=rbind(c(1, 0.8),c(0.8, 1)))
D = Z \%*\% rep(1, 3) + errorTerm[1:10, 2]
Y = D * 1 + errorTerm[1:10, 1]
gl <- group_lasso_iv(Y, D, Z)
model <- fit_tsls(gl)

# Estimate covariance and obtain summary statistics
cov = estimate_covariance(Y, D, Z)
s <- summary_tsls(gl, model$sampler,
                  Sigma_11 = cov[1,1], Sigma_12 = cov[1,2])

# Estimate log density of current optimization samples (using the
# partial log_density function from optimization sampler)
model$sampler$log_density(model$sampler$observed_score_state, s$opt_sample)

# Estimate log density of current optimization samples (directly using the
# log_density function)
log_density(logdens_linear=model$logdens_linear,
            offset=model$opt_offset,
            cond_prec=model$cond_precision,
            score=model$sampler$observed_score_state,
            opt=s$opt_sample,
            active_directions=model$active_directions,
            lagrange=model$lagrange,
            Z=Z)
}
\references{
Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2019). Inference After Selecting
Plausibly Valid Instruments with Application to Mendelian Randomization.

Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2020). Inferring Treatment
Effects After Testing Instrument Strength in Linear Models.
}
\seealso{
\link{log_density_ray} for calculating the log density of normal sample.

\link{log_jacobian} for calculating the log jacobian term of the given
problem.
}
