% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_lasso_iv.R
\name{log_jacobian}
\alias{log_jacobian}
\title{Log Jacobian}
\usage{
log_jacobian(Z, opt, active_directions, lagrange)
}
\arguments{
\item{Z}{The p instruments}

\item{opt}{State for opt variables}

\item{active_directions}{Active directions of vector v (same as opt_linear)}

\item{lagrange}{Lambda term; the weights of penalty}
}
\value{
The log jacobian term.
}
\description{
Calculate the log jacobian term of the given problem. The computed
log jacobian part is used to calculate the log density of
optimization samples.
}
\examples{
# Fit the group lasso optimization model
Z = matrix(rnorm(10*3), nrow = 10, ncol = 3) +
    matrix(replicate(3, matrix(0, nrow = 10, ncol = 1)),
    nrow = 10)
errorTerm = MASS::mvrnorm(n=10, mu=rep(0,2),
            Sigma=rbind(c(1, 0.8),c(0.8, 1)))
D = Z \%*\% rep(1, 3) + errorTerm[1:10, 2]
Y = D * 1 + errorTerm[1:10, 1]
gl <- group_lasso_iv(Y, D, Z)
model <- fit_tsls(gl)

# Estimate covariance and obtain summary statistics
cov = estimate_covariance(Y, D, Z)
s <- summary_tsls(gl, model$sampler,
                  Sigma_11 = cov[1,1], Sigma_12 = cov[1,2])


# Compute the log jacobian part of current optimization samples
log_jacobian(Z=Z,
             opt=s$opt_sample,
             active_directions=model$active_directions,
             lagrange=model$lagrange)
}
\references{
Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2019). Inference After Selecting
Plausibly Valid Instruments with Application to Mendelian Randomization.

Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2020). Inferring Treatment
Effects After Testing Instrument Strength in Linear Models.
}
\seealso{
\link{log_density} for calculating the log density of given samples.
To calculate the log density, a log jacobian part is needed, which
is computed by this function.
}
