% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_lasso_iv_ar.R
\name{naive_inference_ar}
\alias{naive_inference_ar}
\title{Naive Inference for AR Test Statistic}
\usage{
naive_inference_ar(
  Y,
  D,
  Z,
  pass_pre_test,
  parameter = NULL,
  Sigma_11 = 1,
  Sigma_12 = 0.8,
  Sigma_22 = 1,
  compute_intervals = FALSE,
  level = 0.95
)
}
\arguments{
\item{Y}{The outcome.}

\item{D}{The treatment.}

\item{Z}{The p instruments.}

\item{pass_pre_test}{Binary indicator of whether passing the pre-test.}

\item{parameter}{Put beta_0 for null hypothesis.}

\item{Sigma_11}{Variance (i.e. estimate for sigma_11).}

\item{Sigma_12}{Covariance (i.e. estimate for sigma_12).}

\item{Sigma_22}{Variance (i.e. estimate for sigma_22).}

\item{compute_intervals}{Binary indicator of whether to compute confidence intervals.}

\item{level}{Significance level of F-test.}
}
\value{
A naive inference for AR test statistic, which includes components:
\item{ar}{An AR estimate for the treatment effect.}
\item{var_11}{An estimate of variance of AR test statistic.}
\item{var_22}{An estimate of variance of AR test statistic.}
\item{pval}{The p-value of AR test statistic.}
\item{interval}{The confidence interval of AR test statistic.}
}
\description{
Give a simple inference of AR test statistic that does not take
instrument selection/strength test into account.
}
\examples{
# Naive inference for AR test statistic with confidence intervals
# given passing the pre-test
b <- bigaussian_instance(n=10, p=5, s=2, snr=3)
naive_inference_ar(Y=b$Y,
                   D=b$D,
                   Z=b$Z,
                   pass_pre_test=TRUE,
                   compute_intervals=TRUE)

# Naive inference for AR test statistic without confidence intervals
# given passing the pre-test
naive_inference_ar(Y=b$Y, D=b$D, Z=b$Z, pass_pre_test=TRUE)

# Naive inference for AR test statistic given not passing the pre-test
naive_inference_ar(Y=b$Y, D=b$D, Z=b$Z, pass_pre_test=FALSE)
}
\references{
Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2019). Inference After Selecting
Plausibly Valid Instruments with Application to Mendelian Randomization.

Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2020). Inferring Treatment
Effects After Testing Instrument Strength in Linear Models.
}
\seealso{
\link{naive_inference_tsls} that performs a naive inference for TSLS statistic.

\link[stats]{pf} for the distribution function of the F distribution.

\link[stats]{qf} for the quantile function of the F distribution.
}
