% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_lasso_iv.R
\name{naive_inference_tsls}
\alias{naive_inference_tsls}
\title{Naive Inference for TSLS Test Statistic}
\usage{
naive_inference_tsls(
  Y,
  D,
  Z,
  pass_pre_test,
  parameter = NULL,
  Sigma_11 = 1,
  compute_intervals = FALSE,
  level = 0.95
)
}
\arguments{
\item{Y}{The outcome.}

\item{D}{The treatment.}

\item{Z}{The p instruments.}

\item{pass_pre_test}{Binary indicator of whether passing the pre-test.}

\item{parameter}{Put beta_0 for null hypothesis.}

\item{Sigma_11}{Variance (i.e. estimate for sigma_11).}

\item{compute_intervals}{Binary indicator of whether to compute confidence intervals.}

\item{level}{Significance level of F-test.}
}
\value{
A naive inference for TSLS test statistic, which includes components:
\item{tsls}{An TSLS estimate for the treatment effect.}
\item{std}{An estimate of standard error of TSLS test statistic.}
\item{pval}{The p-value of TSLS test statistic.}
\item{interval}{The confidence interval of TSLS test statistic.}
}
\description{
Give a simple inference of TSLS test statistic that does not take
instrument selection/strength test into account.
}
\examples{
# Naive inference for TSLS test statistic with confidence intervals
# given passing the pre-test
b <- bigaussian_instance(n=10, p=5, s=2, snr=3)
naive_inference_tsls(Y=b$Y,
                     D=b$D,
                     Z=b$Z,
                     pass_pre_test=TRUE,
                     compute_intervals=TRUE)

# Naive inference for TSLS test statistic without confidence intervals
# given passing the pre-test
naive_inference_tsls(Y=b$Y, D=b$D, Z=b$Z, pass_pre_test=TRUE)

# Naive inference for TSLS test statistic given not passing the pre-test
naive_inference_tsls(Y=b$Y, D=b$D, Z=b$Z, pass_pre_test=FALSE)
}
\references{
Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2019). Inference After Selecting
Plausibly Valid Instruments with Application to Mendelian Randomization.

Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2020). Inferring Treatment
Effects After Testing Instrument Strength in Linear Models.
}
\seealso{
\link{naive_inference_ar} that performs a naive inference for AR statistic.

\link[stats]{pnorm} for the distribution function of normal distribution.

\link[stats]{qnorm} for the quantile function of normal distribution.
}
