% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{optimization_intervals}
\alias{optimization_intervals}
\title{Optimization Intervals}
\usage{
optimization_intervals(opt_sampling_info, observed, nsample, target_cov = NULL)
}
\arguments{
\item{opt_sampling_info}{A sequence of (opt_sampler, opt_sample, target_cov,
score_cov) objects}

\item{observed}{Observed samples}

\item{nsample}{How large a normal sample}

\item{target_cov}{Covariance of the target (e.g. TSLS)}
}
\value{
An optimization_intervals object, which is a list of:
\item{blahvals}{Repeat opt_samples if necessary}
\item{opt_sampling_info}{A sequence of (opt_sampler, opt_sample, target_cov,
score_cov) objects}
\item{logden}{Total log density of the optimization sampler}
\item{observed}{Observed samples}
\item{target_cov}{Covariance of the target}
\item{normal_sample}{Samples generated from the multivariate normal distribution}
}
\description{
Initialize the optimization_intervals object, which specify the observed
target, target covariance, optimization samples, and optimization sampler.
}
\examples{
# Fit the group lasso optimization model
Z = matrix(rnorm(10*3), nrow = 10, ncol = 3) +
    matrix(replicate(3, matrix(0, nrow = 10, ncol = 1)),
    nrow = 10)
errorTerm = MASS::mvrnorm(n=10, mu=rep(0,2),
            Sigma=rbind(c(1, 0.8),c(0.8, 1)))
D = Z \%*\% rep(1, 3) + errorTerm[1:10, 2]
Y = D * 1 + errorTerm[1:10, 1]
gl <- group_lasso_iv(Y, D, Z)
model <- fit_tsls(gl)

# Estimate covariance and obtain summary statistics
cov = estimate_covariance(Y, D, Z)
s <- summary_tsls(gl, model$sampler,
                  Sigma_11 = cov[1,1], Sigma_12 = cov[1,2])

# Construct the optimization intervals before computing confidence intervals
# for TSLS test statistic
ndraw = nrow(s$opt_sample)
opt_sampling_info = list(sets::tuple(model$sampler,
                                     s$opt_sample,
                                     s$cov_target,
                                     s$cov_target_score))
optimization_intervals(opt_sampling_info,
                       s$observed_target,
                       ndraw)
}
\references{
Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2019). Inference After Selecting
Plausibly Valid Instruments with Application to Mendelian Randomization.

Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2020). Inferring Treatment
Effects After Testing Instrument Strength in Linear Models.
}
\seealso{
\link{log_density} for calculating the log density of a given equation,
including the Jacobian term.

\link[MASS]{mvrnorm} for sampling from multivariate normal distribution.
}
