% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_lasso_iv.R
\name{pre_test}
\alias{pre_test}
\title{Instrument Strength Test}
\usage{
pre_test(Y, D, Z, C0 = NULL)
}
\arguments{
\item{Y}{The outcome.}

\item{D}{The treatment.}

\item{Z}{The p instruments.}

\item{C0}{Threshold of the pre-test.}
}
\value{
TRUE if the pre-test is passed (F-statistic exceeds the threshold)
        and FALSE otherwise.
}
\description{
Simulate a pre-test on instrument strength before inferring the treatment
effect. This method use a F-test to test the instrument strength.
}
\details{
The pre-test regresses the treatment D on the instruments Z to to assess
the instrument strength. The F-test is conducted with the null hypothesis
\eqn{H_0: \gamma^*=0} that there is no association between the treatment and
instruments, so instruments are irrelevant. The alternative hypothesis
\eqn{H_1:\gamma^*\neq 0} implies that there is an association between the treatment
and instruments, so instruments are relevant.
}
\examples{
# Perform a pre-test on instrument strength
Z = matrix(rnorm(10*3), nrow = 10, ncol = 3) +
    matrix(replicate(3, matrix(0, nrow = 10, ncol = 1)),
    nrow = 10)
errorTerm = MASS::mvrnorm(n=10, mu=rep(0,2),
            Sigma=rbind(c(1, 0.8),c(0.8, 1)))
D = Z \%*\% rep(1, 3) + errorTerm[1:10, 2]
Y = D * 1 + errorTerm[1:10, 1]
pre_test(Y, D, Z)
}
\references{
Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2019). Inference After Selecting
Plausibly Valid Instruments with Application to Mendelian Randomization.

Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2020). Inferring Treatment
Effects After Testing Instrument Strength in Linear Models.
}
