% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_from_constraints}
\alias{sample_from_constraints}
\title{Sample From Constraints}
\usage{
sample_from_constraints(
  con,
  Y,
  direction_of_interest = NULL,
  how_often = -1,
  ndraw = 1000,
  burnin = 1000,
  white = FALSE,
  use_constraint_directions = TRUE,
  use_random_directions = TRUE,
  accept_reject_params = NULL
)
}
\arguments{
\item{con}{Constraints}

\item{Y}{Point satisfying the constraint}

\item{direction_of_interest}{Represents which projection is of the most interest}

\item{how_often}{Represents how often would the sampler make a move along "direction_of_interest"
If negative, defaults to ndraw+burnin (so it will never be used).}

\item{ndraw}{Defaults to 1000}

\item{burnin}{Defaults to 1000}

\item{white}{Binary indicator of whether con$covariance is equal to identity}

\item{use_constraint_directions}{Binary indicator of whether using the directions
formed by the constraints as in the Gibbs scheme}

\item{use_random_directions}{Binary indicator of whether using additional
random directions in the Gibbs scheme}

\item{accept_reject_params}{If not NULL should be a tuple (num_trial, min_accept, num_draw).
In this case, we first try num_trial accept-reject samples,
if at least min_accept of them succeed, we just draw num_draw
accept_reject samples.}
}
\value{
Z matrix - sample from the Gaussian distribution conditioned
        on the constraints
}
\description{
Use Gibbs sampler to simulate from "con".
}
\examples{
# Fit the group lasso optimization model
Z = matrix(rnorm(10*3), nrow = 10, ncol = 3) +
    matrix(replicate(3, matrix(0, nrow = 10, ncol = 1)),
    nrow = 10)
errorTerm = MASS::mvrnorm(n=10, mu=rep(0,2),
            Sigma=rbind(c(1, 0.8),c(0.8, 1)))
D = Z \%*\% rep(1, 3) + errorTerm[1:10, 2]
Y = D * 1 + errorTerm[1:10, 1]
gl <- group_lasso_iv(Y, D, Z)
model <- fit_tsls(gl)

# Sample from the box constraint of the optimization problem
sample_from_constraints(con=model$sampler$affine_con,
                        Y=model$sampler$initial_point)
}
\references{
Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2019). Inference After Selecting
Plausibly Valid Instruments with Application to Mendelian Randomization.

Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2020). Inferring Treatment
Effects After Testing Instrument Strength in Linear Models.
}
\seealso{
\link{whiten} for whitening the box constraint.

\link{forward_map} for constructing a forward_map.

\link{accept_reject} for testing random samples.

\link{sample_truncnorm_white} for sampling from a truncated normal
with white constraint.

\link{inverse_map} for constructing an inverse_map.
}
