% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_lasso_iv_ar.R
\name{summary_ar}
\alias{summary_ar}
\title{Summary of Anderson-Rubin Test Statistic}
\usage{
summary_ar(
  gl_ar,
  opt_sampler,
  parameter = NULL,
  Sigma_11 = 1,
  Sigma_12 = 0.8,
  level = 0.95,
  ndraw = 1000,
  burnin = 1000,
  compute_intervals = FALSE,
  compute_power = FALSE,
  beta_alternative = NULL
)
}
\arguments{
\item{gl_ar}{The group lasso object for AR test statistic representing the
penalized convex optimization equation (generated by
group_lasso_iv function)}

\item{opt_sampler}{Sampler to generate opt samples}

\item{parameter}{Put beta_0 for null hypothesis}

\item{Sigma_11}{Variance (i.e. estimate for sigma_11)}

\item{Sigma_12}{Covariance (i.e. estimate for sigma_12)}

\item{level}{Significance level of F-test}

\item{ndraw}{Number of draws/samples after burnin}

\item{burnin}{Number of burnin samples}

\item{compute_intervals}{Binary indicator of whether to compute confidence intervals}

\item{compute_power}{Binary indicator of whether to compute power}

\item{beta_alternative}{A sequence of alternative beta candidates to compute
confidence interval/coverage}
}
\value{
A two-stage least-squares summary object, which is a list of:
\item{observed_target}{Initial AR test statistic}
\item{cov_target}{Covariance of AR test statistic}
\item{cov_target_score}{Covariance of AR score}
\item{two_stage_ls}{TSLS estimator as beta reference for confidence interval}
\item{observed_target_tsls}{Initial TSLS test statistic}
\item{K1}{ZTY matrix of AR test statistic}
\item{K2}{ZTD matrix of AR test statistic}
\item{opt_sample}{Opt samples generated from the box constraint}
\item{pivots}{Pivotal quantities of AR test statistic}
\item{pvalues}{Simulated p-values from truncated normal distribution of AR test statistic}
\item{intervals (if applicable)}{Confidence intervals of AR test statistic}
\item{coverage (if applicable)}{Coverage of AR test statistic}
\item{powers (if applicable}{Powers of AR test statistic}
}
\description{
Give summary statistics of hypothesis testing, including pivots, p-values,
and confidence intervals/powers/coverage of TSLS test statistics.
}
\examples{
# Return summary statistics for AR test statistic
Z = matrix(rnorm(10*3), nrow = 10, ncol = 3) +
    matrix(replicate(3, matrix(0, nrow = 10, ncol = 1)),
    nrow = 10)
errorTerm = MASS::mvrnorm(n=10, mu=rep(0,2),
            Sigma=rbind(c(1, 0.8),c(0.8, 1)))
D = Z \%*\% rep(1, 3) + errorTerm[1:10, 2]
Y = D * 1 + errorTerm[1:10, 1]
gl_ar <- group_lasso_iv_ar(Y, D, Z)
model_ar <- fit_ar(gl_ar)

# Compute confidence intervals
summary_ar(gl_ar, model_ar$sampler, compute_intervals=TRUE)
# Compute power
summary_ar(gl_ar, model_ar$sampler, compute_power=TRUE)
# Compute neither confidence intervals nor power
summary_ar(gl_ar, model_ar$sampler)
}
\references{
Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2019). Inference After Selecting
Plausibly Valid Instruments with Application to Mendelian Randomization.

Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2020). Inferring Treatment
Effects After Testing Instrument Strength in Linear Models.
}
\seealso{
\link{summary_tsls} that returns the summary statistics for group lasso
optimization problem with instrumental variables for an TSLS test statistic.

\link{sample_from_constraints} for using Gibbs sampler to simulate
from the box constraint.

\link{coefficient_pvalues_iv_ar} for constructing selective
p-values for AR statistic.

\link{confidence_intervals_iv_ar} for constructing selective confidence
intervals for AR statistic.

\link{confidence_interval_coverage} for testing selective confidence
intervals for AR statistic.
}
