% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_lasso_iv.R
\name{summary_tsls}
\alias{summary_tsls}
\title{Summary of Two-Stage Least Squares}
\usage{
summary_tsls(
  gl,
  opt_sampler,
  parameter = NULL,
  Sigma_11 = 1,
  Sigma_12 = 0.8,
  level = 0.95,
  ndraw = 1000,
  burnin = 1000,
  compute_intervals = TRUE
)
}
\arguments{
\item{gl}{The group lasso object representing the penalized convex
optimization equation (generated by group_lasso_iv function)}

\item{opt_sampler}{Sampler to generate opt samples}

\item{parameter}{Put beta_0 for null hypothesis}

\item{Sigma_11}{Variance (i.e. estimate for sigma_11)}

\item{Sigma_12}{Covariance (i.e. estimate for sigma_12)}

\item{level}{Significance level of F-test}

\item{ndraw}{Number of draws/samples after burnin}

\item{burnin}{Number of burnin samples}

\item{compute_intervals}{Binary indicator of whether to compute confidence intervals}
}
\value{
A two-stage least-squares summary object, which is a list of:
\item{observed_target}{Initial TSLS test statistic.}
\item{cov_target}{Covariance of TSLS.}
\item{cov_target_score}{Covariance of TSLS score.}
\item{opt_sample}{Opt samples generated from the box constraint.}
\item{pivots}{Pivotal quantities of TSLS test statistic.}
\item{pvalues}{Simulated p-values from truncated normal distribution of TSLS test statistic.}
\item{intervals}{Confidence intervals of TSLS test statistic.}
}
\description{
Give summary statistics of hypothesis testing, including pivots, p-values,
and confidence intervals of TSLS test statistics.
}
\details{
Pivot quantities are generated based on the user-specific null hypothesis
\eqn{H_0: \beta^* = \beta_0}, where \eqn{\beta_0} can be specified through
the argument the argument ‘parameter.’ If the ‘parameter’ argument
is not specified, we test the default null hypothesis \eqn{H_0: \beta^* = \beta_0}.

If all observed target statistics are tested against the null hypothesis
\eqn{H_0: \beta^* = 0}, then p-values will be the same as pivot quantities.
Otherwise, p-values are computed separately under the null hypothesis
\eqn{H_0: \beta^* = 0} instead of the user-specified hypothesis \eqn{H_0: \beta^* = \beta_0},
where \eqn{\beta_0 \neq 0}.

Selective confidence intervals are constructed for the observed target if
the user set the argument ‘compute_intervals’ to be TRUE. This argument
is default to TRUE.
}
\examples{
# Fit the group lasso optimization model
Z = matrix(rnorm(10*3), nrow = 10, ncol = 3) +
    matrix(replicate(3, matrix(0, nrow = 10, ncol = 1)),
    nrow = 10)
errorTerm = MASS::mvrnorm(n=10, mu=rep(0,2),
            Sigma=rbind(c(1, 0.8),c(0.8, 1)))
D = Z \%*\% rep(1, 3) + errorTerm[1:10, 2]
Y = D * 1 + errorTerm[1:10, 1]
gl <- group_lasso_iv(Y, D, Z)
model <- fit_tsls(gl)

# Return summary statistics for TSLS test statistic
summary_tsls(gl, model$sampler)
}
\references{
Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2019). Inference After Selecting
Plausibly Valid Instruments with Application to Mendelian Randomization.

Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2020). Inferring Treatment
Effects After Testing Instrument Strength in Linear Models.
}
\seealso{
\link{summary_ar} that returns the summary statistics for group lasso
optimization problem with instrumental variables for an AR test statistic.

\link{sample_from_constraints} that uses Gibbs sampler
to simulate from the box constraint.

\link{coefficient_pvalues_iv} for constructing selective
p-values for TSLS statistic.

\link{confidence_intervals_iv} for constructing selective
confidence intervals for TSLS statistic.
}
