% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{whiten}
\alias{whiten}
\title{Whiten}
\usage{
whiten(con)
}
\arguments{
\item{con}{The box constraint}
}
\value{
A whiten object, which is a list of:
\item{sqrt_cov}{Squared root of the covariance matrix}
\item{sqrt_inv}{Squared root of the inverse covariance matrix}
\item{rowspace}{Row space matrix}
\item{mu}{Mean vector of Gaussian distribution to be truncated}
\item{linear_part}{The linear part, $A$ of the affine constraint}
\item{offset}{The offset part, $b$ of the affine constraint}
\item{dim}{The dimension of the covariance matrix}
\item{rank}{The rank of the covariance matrix}
\item{covariance}{Covariance matrix of Gaussian distribution to be truncated}
\item{mean}{Mean vector of Gaussian distribution to be truncated}
}
\description{
Return a whitened version of constraints in a different basis, and
a change of basis matrix. If con$covariance is rank deficient, the change-of
basis matrix will not be square.
}
\examples{
# Fit the group lasso optimization model
Z = matrix(rnorm(10*3), nrow = 10, ncol = 3) +
    matrix(replicate(3, matrix(0, nrow = 10, ncol = 1)),
    nrow = 10)
errorTerm = MASS::mvrnorm(n=10, mu=rep(0,2),
            Sigma=rbind(c(1, 0.8),c(0.8, 1)))
D = Z \%*\% rep(1, 3) + errorTerm[1:10, 2]
Y = D * 1 + errorTerm[1:10, 1]
gl <- group_lasso_iv(Y, D, Z)
model <- fit_tsls(gl)

# Whiten the box constraint
whiten(model$sampler$affine_con)
}
\references{
Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2019). Inference After Selecting
Plausibly Valid Instruments with Application to Mendelian Randomization.

Bi, Nan & Kang, Hyunseung & Taylor, Jonathan. (2020). Inferring Treatment
Effects After Testing Instrument Strength in Linear Models.
}
\seealso{
\link{covariance_factors} that finds covariance factor, which is a possibly
non-square square-root.

\link{constraints} for constructing a box constraint for the optimization
problem.
}
